pyproject.toml
[tool.poetry]
name = "opencode-smart-router"
version = "1.0.0"
description = "OpenAI-compatible smart routing proxy for OpenCode"
authors = ["opencode-smart-router"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111"
uvicorn = {extras = ["standard"], version = "^0.29"}
httpx = "^0.27"
pyyaml = "^6.0"
pydantic = "^2.7"
python-dotenv = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^8"
pytest-asyncio = "^0.23"
httpx = "^0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
